requirement.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.9
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
Pillow==8.4.0
pycodestyle==2.8.0
python-decouple==3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
